Create pdfattach.py
I wrote this program initially to replace a page in a PDF. This is helpful when you have to sign and send entire document which requires printing whole document, signing it and scanning entire document again. Instead, just print the page that needed to be attached, sign, scan and use the program with -R option to attach to the specified page number. It removes the specified page and adds the new page.

Later on I added another functionality to extract a page from a PDF file. This is useful when you have to extract a page from document and send it for any reason.